Split out abstract base class from ViewerPanel
This will allow to generalize tools like BookmarksEditor for re-use in
BigVolumeViewer.
